Cast pc_next to WORD in run_ctrl
When performing Register.write to pc, the value is cast to WORD (e.g.
0x80000050).
However, in run_ctrl, pc_next is calculated by adding to current pc an immediate
value that is not cast to WORD. In effect, when adding an immediate,
that represents a negative value (e.g. 0xffffffb0), unsigned addition
will be performed in Python, and the result will be out of range of
WORD. Assigning pc_next to pc will be correct, since it is cast, but
wrong value will be passed to logging.