Increment program counter in advance when syscall
This change may be a workaround or a simplification.
In reality, when control returns from kernel after execution of
a syscall, cpu updates the pc. But it knows how to do that, because it
has additional information, for example stored in CSR registers. When
ECALL happens, 'some information is stored somewhere', the syscall is
executed and when return from ECALL happens, that information is used to
determine new value of pc.
For now, this simulator has no aspirations to include such components of
cpu. Change in this commit assumes, that after the ECALL there will be a
RET and pc should point to that RET after the EXC_ECALL exception.