Add MMU using abstract vm model, don't rely on pt
Here we introduce approach in which the cpu does not rely on the
implementation of the page table.
MMU has a page table attribute. But the only thing MMU assumes about it
is that it has a translate method. The implementation will be on the
kernel side.
The control flow of the mem_access method is partially done. Cases where
permissions are not right are left unimplemented.